refactor parser and formSchedule out of main, fix parsing bug, add comments
