+++ main-site/components/covid-section/CovidSection.styles.ts
