+++ main-site/components/covid-section/CovidSection.tsx
