+++ main-site/components/landing-section/LandingSection.animations.ts
