+++ main-site/components/past-projects-section/PastProjects.animations.ts
