+++ main-site/components/past-projects-section/PastProjectsSection.tsx
