+++ shared-ui/components/404/404.styles.ts
